Publish both deprecated and valid poms of sbt plugins
In sbt 1.9 we fixed the path and pom of sbt plugin published to Maven.
In order to keep the forward compatibility we also publish the deprecated
pom to the deprecated path.

The deprecated and valid poms are different in the pattern of artifact IDs.
For sbt plugins, the artifact IDs of a valid pom contains the sbt cross
version suffix. Thus the directory, the file name and the artifact ID in
the pom are all consistent. Maven is now able to resolve an sbt plugin,
as if it is a ordinary artifact.

Notice though that a newly published sbt plugin can depend on a deprecated
sbt plugin. In which case, Maven will fail to resolve the dependency.
But sbt and Coursier can resolve it successfully.